Added API endpoint with POST and configure React Query
