+++ astro.config.ts
@@ -29,6 +29,8 @@ export default defineConfig({
 		define: {
 			"process.env.PUBLIC_MOVIE_API_KEY": JSON.stringify(process.env.PUBLIC_MOVIE_API_KEY),
 			"process.env.MOVIE_API_KEY": JSON.stringify(process.env.MOVIE_API_KEY),
+			"process.env.PUBLIC_OPENAI_API_KEY": JSON.stringify(process.env.PUBLIC_OPENAI_API_KEY),
+			"process.env.OPENAI_API_KEY": JSON.stringify(process.env.OPENAI_API_KEY),
 		},
 	},
 });
