--- src/features/dashboards/bank-transactions/playground-ai/types.ts
@@ -1,3 +1,0 @@
-export interface ApiResponse {
-  answer: string;
-}
