+++ src/features/dashboards/playground-ai/types.ts
@@ -1,0 +1,7 @@
+export interface OpenAiResponse {
+  content: string;
+}
+
+export interface OpenAiRequest {
+  prompt: string;
+}
