+++ src/pages/dashboards/playground-ai.astro
@@ -1,7 +1,7 @@
 ---
 export const prerender = false;
 import PageLayout from "@/layouts/PageLayout.astro";
-import { PlaygroundAiComponent } from "@/components/portfolio/dashboards/playground-ai";
+import PlaygroundAiAppComponent from "@/components/portfolio/dashboards/playground-ai";
 import ContainerSection from "@/components/common/ContainerSection.astro";
 import HeaderPortfolio from "@/components/portfolio/common/header-portfolio";
 ---
@@ -14,6 +14,6 @@ import HeaderPortfolio from "@/components/portfolio/common/header-portfolio";
 			description={"A responsive user interface for interacting with AI models or chatbots, intuitive and user-friendly. (in progress)"}
 			chipsList={["React", "Typescript", "Tailwind CSS", "Shadcn UI", "Radix UI"]}
 		/>
-		<PlaygroundAiComponent client:only />
+		<PlaygroundAiAppComponent client:only />
 	</ContainerSection>
 </PageLayout>
